+++ components/molecule/thumbnail/src/settings.js
@@ -1,0 +1,22 @@
+export const BASE_CLASS = 'sui-MoleculeThumbnail'
+export const CAPTION_CLASS = `${BASE_CLASS}-caption`
+export const LINK_CLASS = `${BASE_CLASS}-link`
+export const CONTAINER_IMAGE = `${BASE_CLASS}-containerImage`
+
+export const THUMBNAIL_SIZES = {
+  LARGE: 'large',
+  MEDIUM: 'medium',
+  SMALL: 'small',
+  XSMALL: 'xsmall'
+}
+
+export const THUMBNAIL_RATIOS = {
+  '1:1': '1-1',
+  '4:3': '4-3',
+  '16:9': '16-9'
+}
+
+export const THUMBNAIL_SHAPES = {
+  SQUARED: 'squared',
+  CIRCLED: 'circled'
+}